move context and helpers from pages directory
