+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { StudentProvider } from "./context/StudentContext";
+import { StudentProvider } from "../context/StudentContext";
 
 const queryClient = new QueryClient();
